Fix issues found by Coverity Scan (level medium)
Issues in SEP were set to ignore.
Some issues set to ignore based on code intent.

CID1445705 - unchecked cast in details table.
CID1445707 - unchecked fread returned bytes in StarComponent.
CID1445708 - unchecked cast in guider interface.
CID1445711 - unchecked cast in observing list.
CID1445714 - unchecked cast in details dialog.
CID1445719 - unchecked cast in details dialog.
CID1445721 - uninit var hfr in Analyze capture session.
CID1445725 - unchecked cast in EkosLive message.
CID1445726 - dead duplicated code in FITS packer.
CID1445728 - possible division by zero in Scheduler job time estimation.
CID1445729 - uninit var in Linear Focus.
CID1445730 - unchecked cast in EkosLive message.
CID1445734 - uninit vars in SchedulerJob.
CID1445742 - unintended sign extension in FITS star filtering.
CID1445743 - unchecked cast in EkosLive message.
CID1445748 - uninit vars in Meridian Flip test.
CID1445753 - stray semicolon in Capture test helper macro.
CID1445757 - fix if construct (sorry Akarsh).

+++ kstars/fitsviewer/fitsdata.cpp
@@ -1510,7 +1510,7 @@ QFuture<bool> FITSData::findStars(StarAlgorithm algorithm, const QRect &tracking
 
 int FITSData::filterStars(const float innerRadius, const float outerRadius)
 {
-    long const sqDiagonal = this->width() * this->width() / 4 + this->height() * this->height() / 4;
+    long const sqDiagonal = (long) this->width() * (long) this->width() / 4 + (long) this->height() * (long) this->height() / 4;
     long const sqInnerRadius = std::lround(sqDiagonal * innerRadius * innerRadius);
     long const sqOuterRadius = std::lround(sqDiagonal * outerRadius * outerRadius);
